data type for balance changed from float to BigDecimal

+++ card-service/src/main/java/com/idda/project/card_service/dto/response/CardResponse.java
@@ -2,6 +2,7 @@ package com.idda.project.card_service.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.YearMonth;
 
 @Data
@@ -9,7 +10,7 @@ public class CardResponse {
     private Long id;
     private String maskedCardNumber;
     private YearMonth expirationDate;
-    private float balance;
+    private BigDecimal balance;
     private boolean isActive;
 
 }
